+++ tokens/steps.ts
@@ -3,6 +3,14 @@ import type { StepsProps } from '../components/steps'
 
 export type StepContext = {
   uid: number
+  normalColor?: string
+  activeColor?: string
+  left?: number
+  right?: number
+  top?: number
+  bottom?: number
+  width?: number
+  height?: number
 }
 
 export type StepsContext = StepsProps & {