+++ CHANGELOG.md
@@ -1594,3 +1594,73 @@ _2023-09-10_
 
 - `TnActionSheer`新增`maskClosable`参数设置点击遮罩是否关闭
 - `TnPopup`新增`overlay-click`遮罩点击事件
+
+### 1.0.22-beta.1
+
+#### 新增
+
+- `TnSwiper`插槽新增返回`index`参数
+
+### 1.0.22-beta.2
+
+#### 新增
+
+- `TnPicker`新增`allow-confirm-before-scroll-end`处理滚动结束前确认
+- `TnPicker`新增`indicator-height`参数设置指示器高度
+
+### 1.0.22-beta.3
+
+#### 新增
+
+- `TnTabbar`新增`hidden-fixed-tabbar`参数设置隐藏固定底部栏
+
+### 1.0.22-beta.4
+
+#### Fixed
+
+- 修复`TnSteps`横线错位的问题
+
+### 1.0.22-beta.5
+
+#### 新增
+
+- `TnTabbar`新增`bulge-max-size`和`bulge-size-percent`设置中间凸起按钮的尺寸信息
+
+### 1.0.22-beta.6
+### 1.0.22-beta.7
+
+#### 优化
+
+- 优化`TnPicker`无法动态修改多列选择数据的问题
+
+#### 新增
+
+- `TnPicker`新增`immediate-change`参数
+- `TnPicker`新增`pickstart`事件
+- `TnPicker`新增`pickend`事件
+
+### 1.0.22-beta.8
+
+#### 新增
+
+- `TnKeyboard`新增`static`参数设置是否静态显示
+
+### 1.0.22-beta.9
+### 1.0.22-beta.10
+
+#### Fixed
+
+- 修复微信小程序`getSystemInfoSync`问题
+- 修复微信小程序获取胶囊信息出错的问题
+
+### 1.0.22-beta.11
+
+#### 新增
+
+- `TnTabbarItem`新增`content`插槽
+
+### 1.0.23
+
+#### 新增
+
+- `TnWaterFall`新增自定义插槽